Remove unnecessary type conversions + added make command to do the same

+++ roaring_test.go
@@ -372,8 +372,8 @@ func hashTest(t *testing.T, N uint64) {
 			if x+3+gap/2 > MaxUint32 {
 				break
 			}
-			rb1.AddRange(uint64(x), uint64(x+3+gap/2))
-			rb2.AddRange(uint64(x), uint64(x+3+gap/2))
+			rb1.AddRange(x, x+3+gap/2)
+			rb2.AddRange(x, x+3+gap/2)
 		}
 
 		rb1.RunOptimize()
